pyproject.toml
[project]
name = "django-scaffold"
version = "1.0.0"
description = """
基于Django用户组权限管理的Django脚手架，集成JWT认证，celery实现异步任务、以及定时任务的处理，Django Channels处理WebSocket会话。为企业级应用开发提供强大而灵活的基础架构。
"""
license = {text = "MIT"}
authors = [
    {name = "Wytheli",email = "wytheli168@163.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0) ; python_version >= \"3.10\" and python_version < \"4\"",
    "pre-commit (>=4.3.0,<5.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "tencentcloud-sdk-python (>=3.0.1456,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "drf-spectacular[sidecar] (>=0.28.0,<0.29.0)",
    "channels (>=4.3.1,<5.0.0)",
    "channels-redis (>=4.3.0,<5.0.0)",
    "daphne (>=4.2.1,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | venv
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 119
force_single_line = true
